Bütün sayfalar alakalı olduğu role ayrıldı
